Create compositor.cpp
This code sets up a basic QGraphicsScene with a window item (QGraphicsRectItem) that can be moved around using the mouse. The eventFilter function is overridden to handle mouse press and mouse move events on the QGraphicsScene. A QGraphicsDropShadowEffect is applied to the window item to give it a shadow effect.

Although this is just a starting point in this composistor for a foundation, and many contributions will be required.